feat(typography): split into Heading and Text components
