+++ packages/components/src/Heading/Heading.spec.tsx
@@ -1,0 +1,6 @@
+import * as HeadingStoryFile from "./Heading.stories";
+import { generateSnapshots } from "@chanzuckerberg/story-utils";
+
+describe("<Heading />", () => {
+  generateSnapshots(HeadingStoryFile);
+});
