+++ packages/components/src/Text/Text.spec.tsx
@@ -1,0 +1,6 @@
+import * as TextStoryFile from "./Text.stories";
+import { generateSnapshots } from "@chanzuckerberg/story-utils";
+
+describe("<Text />", () => {
+  generateSnapshots(TextStoryFile);
+});
