--- packages/components/src/typography/typography.spec.tsx
@@ -1,6 +1,0 @@
-import * as TypographyStoryFile from "./typography.stories";
-import { generateSnapshots } from "@chanzuckerberg/story-utils";
-
-describe("<Typography />", () => {
-  generateSnapshots(TypographyStoryFile);
-});
